### 2021-07-16 #3503 @whedon
[PRE REVIEW]: mednoise: a python package for medical image processing
**Submitting author:** @ravibandaru-lab (<a href="http://orcid.org/0000-0001-7109-9894">Ravi Bandaru</a>)
**Repository:** <a href="https://github.com/mednoise/mednoise" target ="_blank">https://github.com/mednoise/mednoise</a>
**Version:** v.1.2.4
