### 2021-08-07 #3503 @osorensen
@ethanknights @fredrmag @markushs would any of you be interested in reviewing this submission to the Journal of Open Source Software? It seems to be within your areas of expertise.
