### 2021-08-10 #3503 @markushs
> @ethanknights @fredrmag @markushs would any of you be interested in reviewing this submission to the Journal of Open Source Software? It seems to be within your areas of expertise.

Sorry @osorensen, busy August
