### 2021-09-07 #3597 @osorensen
Dear @ravibandaru-lab
After reading through the paper once more, and testing the submitted software, I am not sure whether the software is within the scope of JOSS. I will therefore pause this review for now, and discuss with the other JOSS editors' whether the submission is within scope.

In particular, whereas the submitted manuscript describes that 'mednoise' contains several algorithms for analyzing medical imaging data, I have not been able to find any demonstrations of such algorithms in the source repository.
