### 2021-09-07 #3597 @ravibandaru-lab
Hello @osorensen!

Hopefully I can clear up some misconceptions/misunderstandings surrounding mednoise.

_mednoise_ is not a tool to **analyze** medical imaging data. I've read over my paper once more to make sure that I didn't make any claims to _mednoise_'s functionality as tool to **analyze** medical imaging data. _mednoise_ is simply a tool to **process** medical images to improve the accuracy of ML and DL models (by removing noise/unneccesary features).

I did, however, make claims and demonstrations to **mednoise**'s algorithms. At the time of submission to this journal, _mednoise_ has three algorithms: hotspot, branch, and manual. The source code for these algorithms are present in my source directory as `branch.py`, `hotspot.py`, and `manual.py`. I've linked them below, as well as their associated testcases below.

- Manual Algorithm Source Code - https://github.com/mednoise/mednoise/blob/main/src/mednoise/manual.py
- Hotspot Algorithm Source Code - https://github.com/mednoise/mednoise/blob/main/src/mednoise/hotspot.py
- Branch Algorithm Source Code - https://github.com/mednoise/mednoise/blob/main/src/mednoise/branch.py

- Manual Algorithm Testcases - https://github.com/mednoise/mednoise/blob/main/tests/tests_manual.py
- Hotspot Algorithm Testcases - https://github.com/mednoise/mednoise/blob/main/tests/tests_hotspot.py
- Branch Algorithm Testcases - https://github.com/mednoise/mednoise/blob/main/tests/tests_branch.py

I've also included a description of the functionality of these algorithms in my documentation here: https://mednoise.github.io/documentation/api_reference/explanation.html. You can also learn more about the algorithms by viewing the API usage guide for them here: https://mednoise.github.io/documentation/api_reference/usage.html

Finally, I've included a sample input and output of the branch algorithm in my paper, which hopefully shows exactly what _mednoise_ does.

Let me know if there is anything else you would like clarification on. Hopefully we can resolve this issue soon.
