### 2021-09-20 #3597 @kthyng
Hi @ravibandaru-lab! Thanks for your submission. Unfortunately, the editorial board has determined that this submission is not currently in scope primarily due to not passing the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort) bar and confusion around what exactly the package does. We do, however, encourage you to improve your work over time and consider submitting it again in the future after more development.

Here is some specific feedback:
* Include explicit references to peer-reviewed publications for the algorithms implemented in the work
* Include examples showing the package being used to show what the software can do in a real life example
* There was a feeling of a disconnect between the scope as described in the paper and docs and in the included functionality -- e.g. the described methods seem concentrated on 2D images like ultrasounds, while the description is medical imaging more broadly ? so even with more development I'd think linking those two would be important !
* There is a lot of duplicated code: Every function in the all.py file also exists (copied) in one of the other files. This matters because it isn't a good coding practice (easy to make errors when you need to update that copied code!), but also it artificially inflates the lines of code in the package.
* The terminology was found to be confusing, which obfuscates the true value/intent of the package.
* It doesn't install because it declares Python standard library components as dependencies in the setup.cfg (they can be left off the dependencies list).
